fix(rag): add api_base (#2289)
# Description

With this pull request, I add a fix that **enables Ollama**. 

Running quivr in a Docker container, Ollama needs to be accessed via
http://host.docker.internal:11434.

This URL is defined as `api_base` of the `ChatLiteLLM` constructor. But
there are calls that are missing `api_base`. This leads to fallback URLs
being used. This is now fixed.

Fixes #2270 #2267 #2217 #2204 

## Checklist before requesting a review

Please delete options that are not relevant.

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged

## Screenshots (if appropriate):

None

Co-authored-by: Niels Garve <niels.garve@experienceone.com>