fix(frontend): send empty sync is not allowed (#2716)
…d files

The syncFiles function in multiple files has been updated to only sync
connections that have selected files. This change improves the
efficiency of the syncing process and ensures that only relevant
connections are synced.

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context.

## Checklist before requesting a review

Please delete options that are not relevant.

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my code
- [ ] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged

## Screenshots (if appropriate):