feat: Add GitHub Actions workflow for running tests on backend/core (#2820)
This commit adds a new GitHub Actions workflow file
`backend-core-tests.yml` to the repository. This workflow is triggered
on push and pull request events, specifically for changes made to the
`backend/core` directory. The workflow runs tests for the backend/core
module using a Tika server as a service. It checks the health of the
Tika server before running the tests. The workflow sets up Python,
installs dependencies using Poetry, and runs the tests using pytest.

This commit is necessary to automate the testing process for the
backend/core module and ensure that the tests are run consistently and
reliably.

Co-authored-by: aminediro <aminedirhoussi@gmail.com>