feat: Remove unused onboarding code (#2901)
Remove unused code related to onboarding functionality. This includes
removing the `OnboardingUpdatableProperties` class and the
`OnboardingStates` class. These classes are no longer used in the
codebase and can be safely deleted.

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context.

## Checklist before requesting a review

Please delete options that are not relevant.

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my code
- [ ] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged

## Screenshots (if appropriate):